+++ lib/samcart_api.rb
@@ -4,6 +4,7 @@ require 'faraday'
 require 'json'
 require 'active_support'
 require 'active_support/core_ext'
+require_relative 'samcart_api/customer'
 
 module SamcartAPI
   VERSION = '0.1.0'