+++ lib/samcart_api/customer.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module SamcartAPI
+  class Customer
+    RESOURCE_PATH = '/customers'
+
+    class << self
+      def find(id)
+        client.get("#{RESOURCE_PATH}/#{id}")
+      end
+
+      def client
+        SamcartAPI::Client.new
+      end
+    end
+  end
+end